[CPPGEN] Moved cppgen from experiments to libs
