+++ libs/CMakeLists.txt
@@ -1,1 +1,2 @@
 add_subdirectory(wayland)
+add_subdirectory(cpp)
