+++ libs/cpp/include/cpp/class.hpp
