+++ libs/cpp/include/cpp/namespace.hpp
