+++ libs/cpp/include/cpp/simple_decl.hpp
